+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- allow `postSubmissionReceipt` if `postSubmissionAction` is `URL`
+
 ## [7.2.0] - 2024-07-10
 
 ### Added